Moved mocked server function to a separated file
