+++ apps/client/app/routes/_layout.templates_.$templateId.tsx
@@ -3,24 +3,11 @@ import { useLoaderData, useNavigate } from "@remix-run/react";
 import { ChevronLeftIcon, PencilIcon } from "@heroicons/react/16/solid";
 import { useState } from "react";
 import parse from "html-react-parser";
-import { MOCKED_TEMPLATES } from "~/mocks/templates";
-import { Template } from "~/models/template.model";
 import CodeView from "~/components/templates/CodeView";
+import { getTemplateById } from "~/services/templates/templates.service";
 
 
 type TemplateViewer = "CONTENT" | "HTML" | "TEXT";
-
-/*
-* Simulates a server request to fetch for a particular template with the specified id
-* */
-async function getTemplateById(id: string) {
-  const result = MOCKED_TEMPLATES.find(template => template.id === id);
-  return new Promise<Template | undefined>((resolve) => {
-    setTimeout(() => {
-      resolve(result);
-      }, 1000)
-  })
-}
 
 export const loader = async ({ params }: LoaderFunctionArgs) => {
   if(params.templateId) {
